pyproject.toml
[project]
name = "uao-neumonia"
version = "0.1.0"
description = "Sistema de Detección de Neumonía con Deep Learning - CLI/Docker"
requires-python = ">=3.12"
dependencies = [
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "pydicom>=3.0.1",
    "reportlab>=4.4.10",
    "tensorflow==2.16.2",
    # Opcional para API
    "flask>=3.1.0",
    "flask-cors>=5.0.0",
    "tensorflow-intel>=2.16.2",
]
